Move rewrite tests into `basic` test suite
