+++ packages/kit/test/apps/basics/src/hooks.js
@@ -1,9 +1,9 @@
 import { browser } from '$app/environment';
 
 const mapping = {
-	'/basic/a': '/basic/b',
-	'/client-only-redirect/a': '/client-only-redirect/b',
-	'/preload-data/a': '/preload-data/b'
+	'/rewrites/basic/a': '/rewrites/basic/b',
+	'/rewrites/client-only-redirect/a': '/rewrites/client-only-redirect/b',
+	'/rewrites/preload-data/a': '/rewrites/preload-data/b'
 };
 
 /** @type {import("@sveltejs/kit").RewriteUrl} */
@@ -13,21 +13,21 @@ export const rewriteUrl = ({ url }) => {
 		return new URL('/', new URL(window.location.href));
 	}
 
-	if (url.pathname === '/error-handling/client-error') {
+	if (url.pathname === '/rewrites/error-handling/client-error') {
 		if (browser) {
 			throw new Error('Client Error');
 		} else {
-			url.pathname = '/error-handling/client-error-rewritten';
+			url.pathname = '/rewrites/error-handling/client-error-rewritten';
 		}
 	}
 
-	if (url.pathname === '/error-handling/server-error') {
+	if (url.pathname === '/rewrites/error-handling/server-error') {
 		throw new Error('Server Error - Should trigger 500 response');
 	}
 
 	url.pathname = mapping[url.pathname] || url.pathname;
 
-	if (url.pathname === '/external/rewritten') {
+	if (url.pathname === '/rewrites/external/rewritten') {
 		return new URL('https://google.com');
 	}
