+++ packages/kit/test/apps/basics/src/routes/rewrites/basic/+page.svelte
@@ -1,0 +1,6 @@
+<script>
+	import { setup } from '../../../../../../setup';
+	setup();
+</script>
+
+<a href="/rewrites/basic/a">Go to url that should be rewritten</a>
