+++ packages/kit/test/apps/basics/src/routes/rewrites/basic/a/+page.svelte
