+++ packages/kit/test/apps/basics/src/routes/rewrites/basic/b/+page.svelte
