+++ packages/kit/test/apps/basics/src/routes/rewrites/client-only-redirect/+page.svelte
@@ -3,4 +3,4 @@
 	setup();
 </script>
 
-<h1>Successfully rewritten</h1>
+<h1>Should be redirected</h1>
