+++ packages/kit/test/apps/basics/src/routes/rewrites/client-only-redirect/a/+page.svelte
@@ -1,0 +1,1 @@
+<h1>Should have been rewritten to <code>/rewrites/client-only-redirect/b</code></h1>
