+++ packages/kit/test/apps/basics/src/routes/rewrites/client-only-redirect/b/+page.svelte
@@ -1,0 +1,6 @@
+<script>
+	import { setup } from '../../../../../../../setup.js';
+	setup();
+</script>
+
+<h1>Successfully rewritten</h1>
