+++ packages/kit/test/apps/basics/src/routes/rewrites/error-handling/+page.svelte
@@ -1,0 +1,8 @@
+<script>
+	import { setup } from '../../../../../../setup';
+
+	setup();
+</script>
+
+<a href="/rewrites/error-handling/client-error" id="client-error">Url with client error</a>
+<a href="/rewrites/error-handling/server-error" id="server-error">Url with server error</a>
