+++ packages/kit/test/apps/basics/src/routes/rewrites/error-handling/client-error-rewritten/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-	import { setup } from '../../../../../../setup';
+	import { setup } from '../../../../../../../setup';
 	setup();
 </script>
