+++ packages/kit/test/apps/basics/src/routes/rewrites/error-handling/client-error/+page.svelte
@@ -1,4 +1,4 @@
 <h1>
 	The rewrite on this page should fail in the browser, causing a full navigation that resolves to
-	<code>/error-handling/client-error-rewritten</code>
+	<code>/rewrites/error-handling/client-error-rewritten</code>
 </h1>
