+++ packages/kit/test/apps/basics/src/routes/rewrites/error-handling/server-error/+page.svelte
