+++ packages/kit/test/apps/basics/src/routes/rewrites/external/+page.svelte
@@ -1,0 +1,8 @@
+<script>
+	import { setup } from '../../../../../../setup';
+
+	setup();
+</script>
+
+<a href="/rewrites/external/rewritten">Go to rewritten page</a>
+<a href="https://expired.badssl.com/" data-test="external-url">External Link</a>
