+++ packages/kit/test/apps/basics/src/routes/rewrites/external/rewritten/+page.svelte
