+++ packages/kit/test/apps/basics/src/routes/rewrites/preload-data/a/+page.js
