+++ packages/kit/test/apps/basics/src/routes/rewrites/preload-data/b/+page.js
