--- packages/kit/test/apps/rewrites/.gitignore
@@ -1,3 +1,0 @@
-/test/errors.json
-!/.env
-/src/routes/routing/symlink-from
