--- packages/kit/test/apps/rewrites/package.json
@@ -1,25 +1,0 @@
-{
-	"name": "test-rewrites",
-	"private": true,
-	"version": "0.0.2-next.0",
-	"scripts": {
-		"dev": "vite dev",
-		"build": "vite build",
-		"preview": "vite preview",
-		"check": "svelte-kit sync && tsc && svelte-check",
-		"test": "node test/setup.js && pnpm test:dev && pnpm test:build",
-		"test:dev": "node -e \"fs.rmSync('test/errors.json', { force: true })\" && cross-env DEV=true playwright test",
-		"test:build": "node -e \"fs.rmSync('test/errors.json', { force: true })\" && cross-env PUBLIC_PRERENDERING=false playwright test"
-	},
-	"devDependencies": {
-		"@sveltejs/kit": "workspace:^",
-		"@sveltejs/vite-plugin-svelte": "^3.0.1",
-		"cross-env": "^7.0.3",
-		"marked": "^11.1.0",
-		"svelte": "^4.2.8",
-		"svelte-check": "^3.6.2",
-		"typescript": "^5.3.3",
-		"vite": "^5.0.8"
-	},
-	"type": "module"
-}
