--- packages/kit/test/apps/rewrites/src/app.d.ts
@@ -1,9 +1,0 @@
-declare global {
-	namespace App {
-		interface Locals {}
-		interface PageState {}
-		interface Platform {}
-	}
-}
-
-export {};
