--- packages/kit/test/apps/rewrites/src/global.d.ts
@@ -1,3 +1,0 @@
-declare global {}
-
-export {};
