--- packages/kit/test/apps/rewrites/src/routes/+page.svelte
@@ -1,1 +1,0 @@
-<h1>SvelteKit Rewrites Test Suite</h1>
