--- packages/kit/test/apps/rewrites/src/routes/basic/+page.svelte
@@ -1,6 +1,0 @@
-<script>
-	import { setup } from '../../../../../setup.js';
-	setup();
-</script>
-
-<a href="/basic/a">Go to url that should be rewritten</a>
