--- packages/kit/test/apps/rewrites/src/routes/client-only-redirect/+page.svelte
@@ -1,6 +1,0 @@
-<script>
-	import { setup } from '../../../../../setup.js';
-	setup();
-</script>
-
-<h1>Should be redirected</h1>
