--- packages/kit/test/apps/rewrites/src/routes/client-only-redirect/a/+page.svelte
@@ -1,1 +1,0 @@
-<h1>Should have been rewritten to <code>/redirect/b</code></h1>
