--- packages/kit/test/apps/rewrites/src/routes/error-handling/+page.svelte
@@ -1,8 +1,0 @@
-<script>
-	import { setup } from '../../../../../setup';
-
-	setup();
-</script>
-
-<a href="/error-handling/client-error" id="client-error">Url with client error</a>
-<a href="/error-handling/server-error" id="server-error">Url with server error</a>
