--- packages/kit/test/apps/rewrites/src/routes/external/+page.svelte
@@ -1,8 +1,0 @@
-<script>
-	import { setup } from '../../../../../setup';
-
-	setup();
-</script>
-
-<a href="/external/rewritten">Go to rewritten page</a>
-<a href="https://expired.badssl.com/" data-test="external-url">External Link</a>
