--- packages/kit/test/apps/rewrites/svelte.config.js
@@ -1,6 +1,0 @@
-/** @type {import('@sveltejs/kit').Config} */
-const config = {
-	kit: {}
-};
-
-export default config;
