--- packages/kit/test/apps/rewrites/test/setup.js
@@ -1,1 +1,0 @@
-export {};
