--- packages/kit/test/apps/rewrites/tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"compilerOptions": {
-		"allowJs": true,
-		"checkJs": true,
-		"esModuleInterop": true,
-		"noEmit": true,
-		"paths": {
-			"@sveltejs/kit": ["../../../types"],
-			"types": ["../../../types/internal"]
-		},
-		"resolveJsonModule": true
-	},
-	"extends": "./.svelte-kit/tsconfig.json"
-}
