--- packages/kit/test/apps/rewrites/vite.config.js
@@ -1,23 +1,0 @@
-import * as path from 'node:path';
-import { sveltekit } from '@sveltejs/kit/vite';
-
-/** @type {import('vite').UserConfig} */
-const config = {
-	build: {
-		minify: false
-	},
-	clearScreen: false,
-	optimizeDeps: {
-		// for CI, we need to explicitly prebundle deps, since
-		// the reload confuses Playwright
-		include: ['cookie', 'marked']
-	},
-	plugins: [sveltekit()],
-	server: {
-		fs: {
-			allow: [path.resolve('../../../src')]
-		}
-	}
-};
-
-export default config;
